+++ thirdeye-core/src/main/java/ai/startree/thirdeye/calcite/dataframe/DataFrameFilterableTable.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package ai.startree.thirdeye.dataframe.calcite;
+package ai.startree.thirdeye.calcite.dataframe;
 
 import ai.startree.thirdeye.spi.dataframe.DataFrame;
 import java.util.List;
