+++ thirdeye-core/src/test/java/ai/startree/thirdeye/calcite/dataframe/DataFrameSchemaTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package ai.startree.thirdeye.dataframe.calcite;
+package ai.startree.thirdeye.calcite.dataframe;
 
 import static ai.startree.thirdeye.detectionpipeline.operator.SqlExecutionOperator.getDataTableFromResultSet;
